requirements.txt
Django==3.2
psycopg2>=2.8
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
request